Set device to global and move sampling related functionality to new file
