+++ src/components/SidePanel/Trigger/TriggerLength.jsx
@@ -38,7 +38,7 @@ import React, { useState } from 'react';
 import Form from 'react-bootstrap/Form';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { triggerLengthUpdate } from '../../../actions/deviceActions';
+import { triggerLengthUpdate } from '../../../actions/triggerActions';
 import { NumberInlineInput, Slider } from '../../../from_pc-nrfconnect-shared';
 import { triggerState } from '../../../reducers/triggerReducer';
